+++ crates/aranya-policy-ifgen-macro/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0](https://github.com/aranya-project/aranya-core/compare/aranya-policy-ifgen-macro-v0.6.0...aranya-policy-ifgen-macro-v0.7.0) - 2026-03-17
+
+### Other
+
+- generate constants for globals ([#593](https://github.com/aranya-project/aranya-core/pull/593))
+
 ## [0.6.0](https://github.com/aranya-project/aranya-core/compare/aranya-policy-ifgen-macro-v0.4.2...aranya-policy-ifgen-macro-v0.6.0) - 2025-11-05
 
 ### Other