+++ crates/aranya-policy-vm/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.20.0](https://github.com/aranya-project/aranya-core/compare/aranya-policy-vm-v0.19.0...aranya-policy-vm-v0.20.0) - 2026-03-17
+
+### Other
+
+- serialize only command values ([#606](https://github.com/aranya-project/aranya-core/pull/606))
+- Close #490: Implement `result` type ([#498](https://github.com/aranya-project/aranya-core/pull/498))
+- split off `ConstValue` and move `Value` to the vm ([#584](https://github.com/aranya-project/aranya-core/pull/584))
+- only compile ffi structs when imported ([#561](https://github.com/aranya-project/aranya-core/pull/561))
+
 ## [0.19.0](https://github.com/aranya-project/aranya-core/compare/aranya-policy-vm-v0.18.0...aranya-policy-vm-v0.19.0) - 2026-02-17
 
 ### Other